+++ modules/AWSPowerShell/Cmdlets/AmazonLambda/Basic/Publish-LMVersion-Cmdlet.cs
@@ -73,7 +73,7 @@ namespace Amazon.PowerShell.Cmdlets.LM
         /// or you can specify Amazon Resource Name (ARN) of the function (for example, <code>arn:aws:lambda:us-west-2:account-id:function:ThumbNail</code>).
         /// AWS Lambda also allows you to specify a partial ARN (for example, <code>account-id:Thumbnail</code>).
         /// Note that the length constraint applies only to the ARN. If you specify only the function
-        /// name, it is limited to 64 character in length. </para>
+        /// name, it is limited to 64 characters in length. </para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true)]