+++ modules/AWSPowerShell/Cmdlets/AmazonLambda/Basic/Update-LMAlias-Cmdlet.cs
@@ -59,7 +59,9 @@ namespace Amazon.PowerShell.Cmdlets.LM
         #region Parameter FunctionName
         /// <summary>
         /// <para>
-        /// <para>The function name for which the alias is created.</para>
+        /// <para>The function name for which the alias is created. Note that the length constraint
+        /// applies only to the ARN. If you specify only the function name, it is limited to 64
+        /// characters in length.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true)]