+++ modules/AWSPowerShell/Cmdlets/AmazonLambda/Basic/Update-LMEventSourceMapping-Cmdlet.cs
@@ -86,7 +86,8 @@ namespace Amazon.PowerShell.Cmdlets.LM
         /// <para>The Lambda function to which you want the stream records sent.</para><para> You can specify a function name (for example, <code>Thumbnail</code>) or you can
         /// specify Amazon Resource Name (ARN) of the function (for example, <code>arn:aws:lambda:us-west-2:account-id:function:ThumbNail</code>).
         /// AWS Lambda also allows you to specify a partial ARN (for example, <code>account-id:Thumbnail</code>).
-        /// </para><para>If you are using versioning, you can also provide a qualified function ARN (ARN that
+        /// Note that the length constraint applies only to the ARN. If you specify only the function
+        /// name, it is limited to 64 characters in length. </para><para>If you are using versioning, you can also provide a qualified function ARN (ARN that
         /// is qualified with function version or alias name as suffix). For more information
         /// about versioning, see <a href="http://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html">AWS
         /// Lambda Function Versioning and Aliases</a></para><para>Note that the length constraint applies only to the ARN. If you specify only the function